Refactor few widgets to dedicated modules
TextButton, CycleHotkeyLabel, ButtonGroup, ToggleHotkeyLabel, List
